+++ compiler/syntax/src/res_ast_debugger.ml
@@ -558,9 +558,6 @@ module SexpAst = struct
             Sexp.list (map_empty ~f:value_binding vbs);
             expression expr;
           ]
-      | Pexp_function cases ->
-        Sexp.list
-          [Sexp.atom "Pexp_function"; Sexp.list (map_empty ~f:case cases)]
       | Pexp_fun (arg_lbl, expr_opt, pat, expr, _) ->
         Sexp.list
           [